BACKEND: Make the rest of the habit operations user-sensitive

+++ backend/routers/habitRouter.js
@@ -28,12 +28,18 @@ habitRouter.get('/', async (req, res, next) => {
   }
 })
 
-habitRouter.get('/:id', async (req, res) => {
-  const habit = await Habit.findById(req.params.id)
-  if (habit) {
-    res.json(habit).status(200)
+habitRouter.get('/:id', async (req, res, next) => {
+  const user = await authorize(req, res, next)
+  if (user) {
+    const habits = await Habit.find({ user: user }).populate('user')
+    const habit = habits.find(h => h.id === req.params.id)
+    if (habit) {
+      res.json(habit).status(200)
+    } else {
+      res.status(404).end()
+    }
   } else {
-    res.status(404).end()
+    res.json({ error: 'invalid token' }).status(401).end()
   }
 })
 
@@ -49,17 +55,39 @@ habitRouter.post('/', async (req, res, next) => {
   }
 })
 
-habitRouter.delete('/:id', async (req, res) => {
-  const habitToDelete = await Habit.findByIdAndDelete(req.params.id)
-  res.status(204).end()
+habitRouter.delete('/:id', async (req, res, next) => {
+  const user = await authorize(req, res, next)
+  if (user) {
+    const habits = await Habit.find({ user: user })
+    const habitToDelete = habits.find(h => h.id === req.params.id)
+    if (habitToDelete) {
+      await Habit.findByIdAndDelete(req.params.id)
+      res.status(204).end()
+    } else {
+      res.json({ error: 'not found' }).status(401).end()
+    }
+  } else {
+    res.json({ error: 'invalid token' }).status(401).end()
+  }
 })
 
-habitRouter.put('/:id', async (req, res) => {
-  try {
-    await Habit.findByIdAndUpdate(req.params.id, req.body)
-    res.status(200).end()
-  } catch (exception) {
-    res.json(exception).status(400).end()
+habitRouter.put('/:id', async (req, res, next) => {
+  const user = await authorize(req, res, next)
+  if (user) {
+    try {
+      const habits = await Habit.find({ user: user })
+      const habitToUpdate = habits.find(h => h.id === req.params.id)
+      if (habitToUpdate) {
+        await Habit.findByIdAndUpdate(req.params.id, req.body)
+        res.status(200).end()
+      } else {
+        res.json({ error: 'not found' })
+      }
+    } catch (exception) {
+      res.json(exception).status(400).end()
+    }
+  } else {
+    res.json({ error: 'invalid token' }).status(401).end()
   }
 })
